+++ src/main/java/id/co/knt/cbt/service/QuestionService.java
@@ -32,5 +32,5 @@ public interface QuestionService {
 	
 	void disabledQG(Long groupId);
 	
-	List<Question> getQuestionByTag(ArrayList<Integer> ids);
+	List<Question> getQuestionByTag(ArrayList<Long> ids);
 }
